% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_openalexR.R
\name{apply_openalexPro}
\alias{apply_openalexPro}
\title{Apply changes to the package \code{openalexR}}
\usage{
apply_openalexPro(which = c("oa_request"))
}
\arguments{
\item{which}{Which function to fix. Can be "oa_request", "api_request", or
both.}
}
\description{
This packagew introduces several changes to the package \code{openalexR}. These can
be applied by specifying the name of the fix. These are at the moment:
\itemize{
\item oa_request: adds option to \code{oa_fetch} to save json files (one per page)
from the call to the OpenAlex API. Changes in the function api_request()
and oa_request() are done.
}
}
