% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_authors.R
\name{extract_authors}
\alias{extract_authors}
\title{Extract authors' institution country codes}
\usage{
extract_authors(
  corpus,
  corpus_authors = "corpus_authors",
  partition = "country"
)
}
\arguments{
\item{corpus}{the corpus in parquet format}

\item{corpus_authors}{parquet dataset of the result. If \code{partition} is `NULL', a file, otherwise a directorty.}

\item{partition}{The column which should be used to partition the table. Hive partitioning is used.
Set to NULL to not partition the table.}
}
\value{
Fully qualified path to the resulting parquet dataset.
}
\description{
This function extracts the institution country codes of authors from given corpus in \code{parquet} format.
}
\examples{
\dontrun{
authors <- data.frame(
  institution_country_code = c("US", "UK", "DE"),
  stringsAsFactors = FALSE
)
extract_authors(authors)
}

}
