% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ids_from_snapshot.R
\name{filter_ids_from_snapshot}
\alias{filter_ids_from_snapshot}
\title{Filter IDs from OA snapshot}
\usage{
filter_ids_from_snapshot(corpus_dir, ids, arrow_dir)
}
\arguments{
\item{corpus_dir}{The directory of the OA snapshot data.}

\item{ids}{either a `id_block` object or vector containing OpenAlex `ids` to filter.}

\item{arrow_dir}{The directory where the filtered IDs will be saved.}
}
\value{
The function returns the path to the directory where the filtered IDs were saved.
}
\description{
This function filters `ids` from an OA snapshot in parquet format (partitioned  by id_block is highly recommended)
and saves the results into the `arrow_dir` partitioned by `id_block`.
}
