% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abbreviated_authors.R
\name{get_abbreviated_authors}
\alias{get_abbreviated_authors}
\title{Abbreviate Authors}
\usage{
get_abbreviated_authors(corpus)
}
\arguments{
\item{corpus}{The corpus. Either a character vector of length 1 with the path to a parquest dataset ot
the corpus itself with the columns `id` and `publication_year` and `authorships`.}
}
\value{
A vector of abbreviated author names and publication years.
}
\description{
This function abbreviates the author names in a given data frame of works.
The output is a list of author names in the format "First Author et al. (Year)" or "Author1 & Author2 (Year)".
}
\examples{
\dontrun{
get_abbreviated_authors(corpus)
}

}
