% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstract.R
\name{get_abstract}
\alias{get_abstract}
\title{Get abstracts from corpus}
\usage{
get_abstract(corpus)
}
\arguments{
\item{corpus}{The corpus. Either a character vector of length 1 with the path to a parquest dataset ot the corpus itself with the columns `id` and `abstract_inverted_index`.}
}
\value{
A named vector with the abstracts of the works.
}
\description{
This function takes a corpus and returns a named vector with the abstracts of the works.
The abstracts are constructed from the abstract inverted index.
}
