% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_tibble.R
\name{json_to_tibble}
\alias{json_to_tibble}
\title{Convert JSON files to a tibble}
\usage{
json_to_tibble(json_dir = file.path("data", "json"))
}
\arguments{
\item{json_dir}{The directory of JSON files returned from \code{json_to_parquet(..., json_dir = "FOLDER")}.}
}
\value{
The tibble of the JSON files.
}
\description{
The function takes a directory of JSON files as written from a call to
\code{json_to_parquet(..., json_dir = "FOLDER")} and converts it to a tibble.
}
\details{
The function uses DuckDB to read the JSON files and to create the
Apache Parquet files. The function creates a DuckDB connection in memory and
readsds the JSON files into DuckDB when needed. Then it creates a SQL query to convert the
JSON files to Apache Parquet files and to copy the result to the specified
directory.
}
\examples{
json_to_tibble(json_dir = "data/json")

}
