% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sql_file.R
\name{load_sql_file}
\alias{load_sql_file}
\title{Load a SQL file}
\usage{
load_sql_file(sql_file = NULL)
}
\arguments{
\item{sql_file}{The path to the SQL file.}
}
\value{
A string containing the SQL code which c an be executed in e.g.
  `DBI::dbExecute(conn, sql)`
}
\description{
Load a SQL file, remove all comments starting with "--" and return the SQL as
a single string.
}
