% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{oa_query}
\alias{oa_query}
\title{Costruct OpenAlex query}
\usage{
oa_query(
  filter = NULL,
  multiple_id = FALSE,
  identifier = NULL,
  entity = if (is.null(identifier)) NULL else id_type(identifier[[1]]),
  options = NULL,
  search = NULL,
  group_by = NULL,
  endpoint = "https://api.openalex.org",
  verbose = FALSE,
  ...
)
}
\description{
Re-export from openalexR::oa_query()
}
