% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_request.R
\name{oa_request}
\alias{oa_request}
\title{\code{openalexR::oa_request()} with additional argument}
\usage{
oa_request(
  query_url,
  per_page = 200,
  paging = "cursor",
  pages = NULL,
  count_only = FALSE,
  mailto = oa_email(),
  api_key = oa_apikey(),
  verbose = FALSE,
  json_dir = NULL
)
}
\arguments{
\item{json_dir}{directory where the JSON files are saved. Default is NULL}
}
\value{
If \code{json_dir} is \code{NULL}, the return value from call to \code{openalexR::oa_request()},
otherwise the complete path to the expanded and normalized \code{json_dir}.
}
\description{
This function adds one argument to \code{openalexR::oa_request()}, namely \code{json_dir}.
When specified, all return values from OpenAlex will be saved as jaon files in
that directory and the return value is the directory of the json files.
}
\details{
For the documentation please see \code{openalexR::oa_request()}
}
