% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_api_key.R
\name{opt_api_key}
\alias{opt_api_key}
\title{Get API key for OpenAlex API}
\usage{
opt_api_key(api_key)
}
\arguments{
\item{api_key}{character vector or NULL. If specified, value to assign to
the api key option. Default is `NULL`.}
}
\value{
The API key, if `api_key` is not specified the current one, otherwise the old one.
}
\description{
Get API key for OpenAlex API
}
