% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_filter_names.R
\name{opt_filter_names}
\alias{opt_filter_names}
\title{Get available filter names from OpenAlex API}
\usage{
opt_filter_names(update = FALSE)
}
\arguments{
\item{update}{logical. If `TRUE` update the existing value. Default is `FALSE`.}
}
\value{
A character vector of available filter names
}
\description{
Get available filter names from OpenAlex API
}
