% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_select_fields.R
\name{opt_select_fields}
\alias{opt_select_fields}
\title{Get available select fields from OpenAlex API}
\usage{
opt_select_fields(update = FALSE)
}
\arguments{
\item{update}{logical. If `TRUE` update the existing value. Default is `FALSE`.}
}
\value{
A character vector of available select fields
}
\description{
Get available select fields from OpenAlex API
}
