% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_api_key.R
\name{pro_api_key}
\alias{pro_api_key}
\title{Retrieve the OpenAlex Pro API key}
\usage{
pro_api_key()
}
\value{
A character string containing the API key, or \code{NULL} if no key
could be found.
}
\description{
Retrieves the OpenAlex Pro API key from one of several locations,
checked in the following order:
}
\details{
\enumerate{
  \item The R option \code{openalexPro$api_key}
  \item The environment variable \code{openalexPro.api_key}
  \item The system keyring via the \pkg{keyring} package
        (only if the package \pkg{keyring} is installed)
}

If no API key is found, \code{NULL} or an empty string may be returned,
depending on the environment variable state.
}
\examples{
\dontrun{
pro_api_key()

options(openalexPro = list(api_key = "my-key"))
pro_api_key()
}

}
\seealso{
\code{\link[base:options]{options}},
\code{\link[base:Sys.getenv]{Sys.getenv}}
}
