% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_request_json_to_parquet.R
\name{pro_request_json_to_parquet}
\alias{pro_request_json_to_parquet}
\title{Convert JSON files to Apache Parquet files}
\usage{
pro_request_json_to_parquet(
  source_dir = NULL,
  corpus = tempfile(fileext = ".corpus"),
  verbose = FALSE
)
}
\arguments{
\item{source_dir}{The directory of JSON files returned from \code{pro_request(..., json_dir = "FOLDER")}.}

\item{corpus}{parquet dataset; default: temporary directory.}

\item{verbose}{Logical indicating whether to show a verbose information. Defaults to \code{FALSE}}
}
\value{
The function does not return anything, but it creates a directory with
Apache Parquet files.
}
\description{
The function takes a directory of JSON files as written from a call to \code{pro_request(..., source_dir = "FOLDER")}
and converts it to a Apache Parquet dataset partitiond by the page.
}
\details{
The function uses DuckDB to read the JSON files and to create the
Apache Parquet files. The function creates a DuckDB connection in memory and
readsds the JSON files into DuckDB when needed. Then it creates a SQL query to convert the
JSON files to Apache Parquet files and to copy the result to the specified
directory.
}
\examples{
\dontrun{
source_to_parquet(source_dir = "json", source_type = "snapshot", corpus = "arrow")
}
}
