% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_request_jsonl.R
\name{pro_request_jsonl}
\alias{pro_request_jsonl}
\title{Convert JSON files to Apache Parquet files}
\usage{
pro_request_jsonl(
  input_json = NULL,
  output = NULL,
  add_columns = list(),
  overwrite = FALSE,
  verbose = TRUE,
  delete_input = FALSE,
  jq_path = "jq"
)
}
\arguments{
\item{input_json}{The directory of JSON files returned from \code{pro_request(..., json_dir = "FOLDER")}.}

\item{output}{output directory for the jsonl files as created by calls to `jq_execute().}

\item{add_columns}{List of additional fields to be added to the output. They nave to be provided as a
named list, e./g. \code{list(column_1 = "value_1", column_2 = 2)}. Only Scalar values are supported.}

\item{overwrite}{Logical indicating whether to overwrite \code{poutput_jsonl} and \code{output_nn}.}

\item{verbose}{Logical indicating whether to show a verbose information. Defaults to \code{TRUE}}

\item{delete_input}{Determines if the \code{input_json} should be deleted afterwards. Defaults to \code{FALSE}.}

\item{jq_path}{Path to the jq executable (default: "jq")}
}
\value{
The function does returns the output invisibly.
}
\description{
The function takes a directory of JSON files as written from a call to \code{pro_request(..., output = "FOLDER")}
and is preparing the json files to be processed further using DuckDB. See
}
\details{
See \link[jq_execute]{jq_execute} or the vignette\link[vignette:jq]{Expanding Abstracts and Adding Citations Using jq}
for more information on the conversion of the JSON files.

The function uses DuckDB to read the JSON files and to create the
Apache Parquet files. The function creates a DuckDB connection in memory and
readsds the JSON files into DuckDB when needed. Then it creates a SQL query to convert the
JSON files to Apache Parquet files and to copy the result to the specified
directory.
}
\examples{
\dontrun{
source_to_parquet(input_json = "json", source_type = "snapshot", poutput_jsonl = "arrow")
}
}
