% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_snowball.R
\name{pro_snowball}
\alias{pro_snowball}
\title{A function to perform a snowball search
and convert the result to a tibble/data frame.}
\usage{
pro_snowball(
  identifier = NULL,
  doi = NULL,
  output = tempfile(fileext = ".snowball"),
  verbose = FALSE
)
}
\arguments{
\item{identifier}{Character vector of openalex identifiers.}

\item{doi}{Character vector of dois.}

\item{output}{parquet dataset; default: temporary directory.}

\item{verbose}{Logical indicating whether to show a verbose information. Defaults to \code{FALSE}}
}
\value{
A list containing 2 elements:
\itemize{
\item nodes: dataframe with publication records.
The last column \code{oa_input} indicates whether the work was
one of the input \code{identifier}(s).
\item edges: publication link dataframe of 2 columns \verb{from, to}
such that a row \verb{A, B} means A -> B means A cites B.
In bibliometrics, the "citation action" comes from A to B.
}
}
\description{
A function to perform a snowball search
and convert the result to a tibble/data frame.
}
\examples{
\dontrun{

snowball_docs <- pro_snowball(
  identifier = c("W2741809807", "W2755950973"),
  citing_params = list(from_publication_date = "2022-01-01"),
  cited_by_params = list(),
  verbose = TRUE
)

# Identical to above, but searches using paper DOIs
snowball_docs_doi <- oa_snowball(
  doi = c("10.1016/j.joi.2017.08.007", "10.7717/peerj.4375"),
  citing_params = list(from_publication_date = "2022-01-01"),
  cited_by_params = list(),
  verbose = TRUE
)
}
}
