% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{oa_apikey}
\alias{oa_email}
\alias{oa_query}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{openalexR}{\code{\link[openalexR]{oa_apikey}}, \code{\link[openalexR]{oa_email}}, \code{\link[openalexR]{oa_query}}}
}}

