% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_openalexR.R
\name{unapply_openalexPro}
\alias{unapply_openalexPro}
\title{Revert **all** changes to the package `openalexR` done by `apply_openalexPro()`}
\usage{
unapply_openalexPro()
}
\description{
This function undoes the changes introduced by \code{\link{apply_openalexPro}}.
It unloads the namespace and reloads the package.
}
\examples{
\dontrun{
apply_openalexPro()
unapply_openalexPro()
}
}
