% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corpus_index.R
\name{build_corpus_index}
\alias{build_corpus_index}
\title{Build a Parquet ID-lookup index}
\usage{
build_corpus_index(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
**Moved to the \pkg{openalexSnapshot} package.**

This function has been removed from \pkg{openalexPro}.
Please install the \pkg{openalexSnapshot} package and call
`openalexSnapshot::build_corpus_index()` instead.
}
\seealso{
\url{https://github.com/rkrug/openalexSnapshot}
}
