% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_block.R
\name{id_block}
\alias{id_block}
\title{Create a table with id blocks}
\usage{
id_block(ids = NULL)
}
\arguments{
\item{ids}{A vector of IDs}
}
\value{
A tibble with additional class \code{id_block} with the columns 'id\verb{, }long_id\verb{, }short_id\code{and}id_block`.
}
\description{
This function takes a vector of OpenAlex \code{ids}, and determines the
\itemize{
\item short_id
\item long_id
\item id_block
}
}
\examples{
id_block(ids = c("https://openalex.org/W2582743722", "W2582756"))

}
