% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_by_id.R
\name{lookup_by_id}
\alias{lookup_by_id}
\title{Look up records by OpenAlex ID}
\usage{
lookup_by_id(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
**Moved to the \pkg{openalexSnapshot} package.**

This function has been removed from \pkg{openalexPro}.
Please install the \pkg{openalexSnapshot} package and call
`openalexSnapshot::lookup_by_id()` instead.
}
\seealso{
\url{https://github.com/rkrug/openalexSnapshot}
}
