% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_request.R
\name{oa_request}
\alias{oa_request}
\title{\code{openalexR::oa_request()} with additional argument}
\usage{
oa_request(
  query_url,
  per_page = 200,
  paging = "cursor",
  pages = NULL,
  count_only = FALSE,
  mailto = oa_email(),
  api_key = oa_apikey(),
  verbose = FALSE,
  json_dir = NULL
)
}
\arguments{
\item{query_url}{The URL of the API query.}

\item{per_page}{The number of items to be returned per page. Defaults to 200.}

\item{paging}{The type of paging. Possible values are "page" and "cursor".}

\item{pages}{The number of pages to be downloaded. If \code{NULL}, all pages will be downloaded.}

\item{count_only}{Logical indicating whether to return only the count of the items returned by the query.}

\item{mailto}{The email address of the user. See \code{openalexR::oa_email()}.}

\item{api_key}{The API key of the user. See \code{openalexR::oa_apikey()}.}

\item{verbose}{Logical indicating whether to show a progress bar.}

\item{json_dir}{directory where the JSON files are saved. Default is NULL}
}
\value{
If \code{json_dir} is \code{NULL}, the return value from call to \code{openalexR::oa_request()},
otherwise the complete path to the expanded and normalized \code{json_dir}.
}
\description{
This function adds one argument to \code{openalexR::oa_request()}, namely \code{json_dir}.
When specified, all return values from OpenAlex will be saved as jaon files in
that directory and the return value is the directory of the json files.
}
\details{
For the documentation please see \code{openalexR::oa_request()}
}
