% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_request.R
\name{pro_request}
\alias{pro_request}
\title{Fetch works from OpenAlex}
\usage{
pro_request(
  query_url,
  pages = 1e+05,
  output = NULL,
  overwrite = FALSE,
  mailto = Sys.getenv("openalexPro.email"),
  api_key = Sys.getenv("openalexPro.apikey"),
  workers = 1,
  verbose = FALSE,
  progress = TRUE,
  count_only = FALSE,
  error_log = NULL
)
}
\arguments{
\item{query_url}{The URL of the API query or a list of URLs returned from \code{pro_query()}.}

\item{pages}{The number of pages to be downloaded. The default is set to
10000, which would be 2,000,000 works. It is recommended to not increase it
beyond 100000 due to server load and to use the snapshot instead. If \code{NULL},
all pages will be downloaded. Default: 100000.}

\item{output}{directory where the JSON files are saved. Default is a
temporary directory. Needs to be specified.}

\item{overwrite}{Logical. If \code{TRUE}, \code{output} will be deleted if it already
exists.}

\item{mailto}{The email address of the user.}

\item{api_key}{The API key of the user.}

\item{workers}{Number of parallel workers to use if \code{query_url} is a list. Defaults to 1.}

\item{verbose}{Logical indicating whether to show verbose messages.}

\item{progress}{Logical indicating whether to show a progress bar. Default \code{TRUE}.}

\item{count_only}{return count only as a named numeric vector or list.}

\item{error_log}{location of error log of API calls. (default: \code{NULL} (none)).}
}
\value{
If \code{count_only} is \code{FALSE} (the default) the complete path to the expanded and
normalized \code{output}. If \code{count_only} is \code{TRUE}, a named numeric vector with the count
of the works from the specified query_url(s).
}
\description{
All returned values from OpenAlex will be saved as
json files in the \code{output} directory and the return value is the directory of the
json files.
}
\details{
If query_url is a list, the function is called for each element of the list in parallel
using a maximum of \code{workers} parallel R sessions. The results from the individual URLs
in the list are returned in a folder named after the names of the list elements in the
\code{output} folder.

When starting the download, a file \verb{00_in.progress} which is deleted upon completion.
}
