% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_request.R
\name{pro_request}
\alias{pro_request}
\title{\code{openalexR::oa_request()} with additional argument}
\usage{
pro_request(
  query_url,
  pages = 1000,
  output = NULL,
  overwrite = FALSE,
  mailto = oap_mail(),
  api_key = oap_apikey,
  workers = 1,
  verbose = FALSE,
  progress = TRUE,
  count_only = FALSE,
  error_log = NULL
)
}
\arguments{
\item{query_url}{The URL of the API query or a list of URLs returned from \code{pro_query()}.}

\item{pages}{The number of pages to be downloaded. The default is set to
1000, which would be 2,000,000 works. It is recommended to not increase it
beyond 1000 due to server load and to use the snapshot instead. If \code{NULL},
all pages will be downloaded. Default: 1000.}

\item{output}{directory where the JSON files are saved. Default is a
temporary directory. If \code{NULL}, the return value from call to
\code{openalexR::oa_request()} with all the arguments is returned}

\item{overwrite}{Logical. If \code{TRUE}, \code{output} will be deleted if it already
exists.}

\item{mailto}{The email address of the user. See \code{oap_mail()}.}

\item{api_key}{The API key of the user. See \code{oap_apikey()}.}

\item{workers}{Number of parallel workers to use if \code{query_url} is a list. Defaults to 1.}

\item{verbose}{Logical indicating whether to show verbose messages.}

\item{progress}{Logical default \code{TRUE} indicating whether to show a progress
bar.}

\item{count_only}{return count only as a named numeric vector or list.}

\item{error_log}{location of error log of API calls. (default: \code{NULL} (none)).}
}
\value{
If \code{count_only} is \code{FALSE} (the default) the complete path to the expanded and
normalized \code{output}. If \code{count_only} is \code{TRUE}, a named numeric vector with the count
of the works from the specified query_url(s).
}
\description{
This function adds one argument to \code{openalexR::oa_request()}, namely
\code{output}. When specified, all return values from OpenAlex will be saved as
json files in that directory and the return value is the directory of the
json files.
}
\details{
For the documentation please see \code{openalexR::oa_request()}
If query_url is a list, the function is called for each element of the list in parallel
using a maximum of \code{workers} parallel R sessions. The results from the individual URLs
in the list are returned in a folder named after the names of the list elements in the
\code{output} folder.

Nested progress bars:
\itemize{
\item outer bar: list elements (queries)
\item inner bar: pages per query
}

If the \code{progressr} package is not available or \code{progress = FALSE}, the function
falls back to a simple \code{txtProgressBar} (per-query) or no progress at all.
}
