% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_snowball.R
\name{pro_snowball}
\alias{pro_snowball}
\title{A function to perform a snowball search and convert the result to a
tibble/data frame.}
\usage{
pro_snowball(
  identifier = NULL,
  doi = NULL,
  output = tempfile(fileext = ".snowball"),
  verbose = FALSE
)
}
\arguments{
\item{identifier}{Character vector of openalex identifiers.}

\item{doi}{Character vector of dois.}

\item{output}{parquet dataset; default: temporary directory.}

\item{verbose}{Logical indicating whether to show a verbose information.
Defaults to \code{FALSE}}
}
\value{
The folder of the results containing multiple subfolders.
}
\description{
A function to perform a snowball search and convert the result to a
tibble/data frame.
}
