% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_snowball_get_nodes.R
\name{pro_snowball_get_nodes}
\alias{pro_snowball_get_nodes}
\title{A function to get the nodes for a snowball search}
\usage{
pro_snowball_get_nodes(
  identifier = NULL,
  doi = NULL,
  output = tempfile(fileext = ".snowball"),
  verbose = FALSE
)
}
\arguments{
\item{identifier}{Character vector of openalex identifiers.}

\item{doi}{Character vector of dois.}

\item{output}{parquet dataset; default: temporary directory.}

\item{verbose}{Logical indicating whether to show a verbose information.
Defaults to \code{FALSE}}
}
\value{
Path to the nodes parquet dataset
}
\description{
A function to get the nodes for a snowball search
}
\examples{
\dontrun{

snowball_docs <- pro_snowball( identifier = c("W2741809807", "W2755950973"),
citing_params = list(from_publication_date = "2022-01-01"), cited_by_params =
list(), verbose = TRUE )

# Identical to above, but searches using paper DOIs snowball_docs_doi <-
oa_snowball( doi = c("10.1016/j.joi.2017.08.007", "10.7717/peerj.4375"),
citing_params = list(from_publication_date = "2022-01-01"), cited_by_params =
list(), verbose = TRUE ) }
}
