% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_validate_credentials.R
\name{pro_validate_credentials}
\alias{pro_validate_credentials}
\title{Validate OpenAlex credentials}
\usage{
pro_validate_credentials(
  api_key = Sys.getenv("openalexPro.apikey"),
  show_credentials = FALSE
)
}
\arguments{
\item{api_key}{API key to validate (character string) or `NULL`.}

\item{show_credentials}{shows the api_key using `message()`. USE WITH CAUTION!}
}
\value{
TRUE if credentials work, FALSE otherwise
}
\description{
Makes a minimal API request to verify that the api_key is valid.
}
