% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpus.R
\name{read_corpus}
\alias{read_corpus}
\title{Read corpus from Parquet Dataset}
\usage{
read_corpus(corpus, return_data = FALSE, comp_mode = FALSE)
}
\arguments{
\item{corpus}{The directory of the Parquet files.}

\item{return_data}{Logical indicating whether to return an Arrow Dataset
(default) or a \code{tibble}.}

\item{comp_mode}{Compatibility mode with \code{openalexR} return values from  \code{openalexR::fetch( output = "tibble")} (default: \code{FALSE}).
If \verb{TRUE,}, the format of the returned values is identical to \code{openalexR::fetch( output = "tibble")},
if \code{FALSE}, the format may be functionally identical, but not structurally.}
}
\value{
An Arrow Dataset or a \code{tibble}.
}
\description{
This function reads a corpus in Apache Parquet format and returns an Arrow
Dataset or a \code{tibble}.
}
