% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snowball.R
\name{read_snowball}
\alias{read_snowball}
\title{Read snowball from Parquet Dataset}
\usage{
read_snowball(
  snowball = NULL,
  edge_type = c("core", "extended"),
  return_data = FALSE
)
}
\arguments{
\item{snowball}{The directory of the Parquet files as poppulater by \code{pro_snowball()}.}

\item{edge_type}{description}

\item{return_data}{Logical indicating whether to return an
\code{ArrowObject} representing the corpus (default) or a
\code{tibble} containing the whole corpus shou,d be returned.}
}
\value{
A list containing two elements: nodes and edges, which are either
\code{ArrowObject} representing the corpus or \code{tibble}s containing the data.
}
\description{
This function reads a snowball from Apache Parquet format and returns a list containing
nodes and edges, which can be either Arrow Datasets or \code{tibble}s.
}
