% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_to_parquet.R
\name{snapshot_to_parquet}
\alias{snapshot_to_parquet}
\title{Convert OA snapshot to Parquet format}
\usage{
snapshot_to_parquet(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
**Moved to the \pkg{openalexSnapshot} package.**

This function has been removed from \pkg{openalexPro}.
Please install the \pkg{openalexSnapshot} package and call
`openalexSnapshot::snapshot_to_parquet()` instead.
}
\seealso{
\url{https://github.com/rkrug/openalexSnapshot}
}
